arch/arm/src/sama5/sam_sdmmc.c: Fix a type mismatch

+++ arch/arm/src/sama5/sam_sdmmc.c
@@ -328,7 +328,7 @@ static void sam_blocksetup(FAR struct sdio_dev_s *dev,
 static int  sam_recvsetup(FAR struct sdio_dev_s *dev, FAR uint8_t *buffer,
               size_t nbytes);
 static int  sam_sendsetup(FAR struct sdio_dev_s *dev,
-              FAR const uint8_t *buffer, uint32_t nbytes);
+              FAR const uint8_t *buffer, size_t nbytes);
 #endif
 
 static int  sam_cancel(FAR struct sdio_dev_s *dev);
